+++ examples/live/multi_venue_ema_cross_stop_entry_with_trail.py
@@ -59,6 +59,8 @@ config = {
         "port": 6379,
     },
 
+    "risk": {},
+
     "strategy": {
         "load_state": True,  # Strategy state is loaded from the database on start
         "save_state": True,  # Strategy state is saved to the database on shutdown