+++ tests/unit_tests/common/test_common_providers.py
@@ -14,6 +14,7 @@
 # -------------------------------------------------------------------------------------------------
 
 import asyncio
+
 import pytest
 
 from nautilus_trader.common.providers import InstrumentProvider